requirements.txt
alembic
fastapi
fastapi-mcp
"git+https://github.com/sparfenyuk/mcp-proxy" #mcp-proxy
passlib[bcrypt]
psycopg2
pydantic
pyjwt
python-decouple
python-multipart
sqlmodel
uvicorn